+++ R/export_template.R
@@ -2,12 +2,22 @@
 #'
 #' @param src_lang Source language -- "en" for english is the only supported one.
 #'
-#' @return The path to the template file (class: character).
+#' @return The path to the template file (class: character). An error is raised
+#'   if no template is shipped for `src_lang`.
 #'
 #' @examples
 #' export.template()
 #'
 #' @export
 export.template <- function(src_lang = "en") {
-  system.file("templates/en_en.csv", package = "tr.iatgen")
+  template <- pkg.file(file.path("templates", paste0(src_lang, "_", src_lang, ".csv")))
+
+  if (!nzchar(template)) {
+    stop(
+      "No translation template available for `src_lang` \"", src_lang,
+      "\". Only \"en\" is currently supported."
+    )
+  }
+
+  template
 }